Add initial TCP echo code
